feat(shader): improve ShaderProperty class with enhanced methods and formatting
- Refactored constructor and getter methods for better readability.
- Added static methods for managing shader property IDs and names.
- Improved code formatting and consistency across the ShaderProperty class.

feat(engine): enhance demo scene creation and engine initialization

- Updated createDemoScene function for better structure and readability.
- Improved createDemoEngine function with better handling of options and resizing.
- Added error handling for missing container elements in startDemo function.

feat(math): add new matrix operations for transformations

- Implemented orthographic projection matrix creation.
- Added rotation methods for arbitrary axes and standard axes (X, Y, Z).
- Introduced scaling and lookAt methods for matrix transformations.
- Enhanced matrix manipulation with transpose and Euler angle support.
