+++ packages/core/src/animation/AnimationClip.ts
@@ -6,15 +6,15 @@ import { ReferResource } from '../base/ReferResource';
  */
 export interface Keyframe<T> {
   /** 时间(秒) */
-  time: number;
+  time: number,
   /** 值 */
-  value: T;
+  value: T,
   /** 插值方式 */
-  interpolation?: InterpolationType;
+  interpolation?: InterpolationType,
   /** 输入切线(仅用于曲线插值) */
-  inTangent?: number | number[];
+  inTangent?: number | number[],
   /** 输出切线(仅用于曲线插值) */
-  outTangent?: number | number[];
+  outTangent?: number | number[],
 }
 
 /**
@@ -22,13 +22,13 @@ export interface Keyframe<T> {
  */
 export interface AnimationTrack<T> {
   /** 通道名称 */
-  name: string;
+  name: string,
   /** 目标属性路径 */
-  path: string;
+  path: string,
   /** 关键帧数据 */
-  keyframes: Keyframe<T>[];
+  keyframes: Keyframe<T>[],
   /** 默认插值类型 */
-  interpolation: InterpolationType;
+  interpolation: InterpolationType,
 }
 
 /**
@@ -51,7 +51,7 @@ export class AnimationClip extends ReferResource {
    * 创建动画片段
    * @param name 片段名称
    */
-  constructor(name: string) {
+  constructor (name: string) {
     super();
     this.name = name;
   }
@@ -59,7 +59,7 @@ export class AnimationClip extends ReferResource {
   /**
    * 获取片段名称
    */
-  getName(): string {
+  getName (): string {
     return this.name;
   }
 
@@ -77,19 +77,22 @@ export class AnimationClip extends ReferResource {
    * @param nameOrIndex 通道名称或索引
    * @returns 是否成功移除
    */
-  removeTrack(nameOrIndex: string | number): boolean {
+  removeTrack (nameOrIndex: string | number): boolean {
     if (typeof nameOrIndex === 'string') {
       const index = this.tracks.findIndex(track => track.name === nameOrIndex);
+
       if (index === -1) {return false;}
-      
+
       this.tracks.splice(index, 1);
       this.updateDuration();
+
       return true;
     } else {
       if (nameOrIndex < 0 || nameOrIndex >= this.tracks.length) {return false;}
-      
+
       this.tracks.splice(nameOrIndex, 1);
       this.updateDuration();
+
       return true;
     }
   }
@@ -110,26 +113,27 @@ export class AnimationClip extends ReferResource {
   /**
    * 获取所有通道
    */
-  getAllTracks(): AnimationTrack<any>[] {
+  getAllTracks (): AnimationTrack<any>[] {
     return [...this.tracks];
   }
 
   /**
    * 获取通道数量
    */
-  getTrackCount(): number {
+  getTrackCount (): number {
     return this.tracks.length;
   }
 
   /**
    * 更新动画时长
    */
-  private updateDuration(): void {
+  private updateDuration (): void {
     this.duration = 0;
-    
+
     for (const track of this.tracks) {
       if (track.keyframes.length > 0) {
         const lastKeyframe = track.keyframes[track.keyframes.length - 1];
+
         this.duration = Math.max(this.duration, lastKeyframe.time);
       }
     }
@@ -138,7 +142,7 @@ export class AnimationClip extends ReferResource {
   /**
    * 获取动画时长
    */
-  getDuration(): number {
+  getDuration (): number {
     return this.duration;
   }
 
@@ -146,14 +150,14 @@ export class AnimationClip extends ReferResource {
    * 设置是否循环
    * @param loop 是否循环
    */
-  setLoop(loop: boolean): void {
+  setLoop (loop: boolean): void {
     this.loop = loop;
   }
 
   /**
    * 是否循环
    */
-  isLoop(): boolean {
+  isLoop (): boolean {
     return this.loop;
   }
 
@@ -161,14 +165,14 @@ export class AnimationClip extends ReferResource {
    * 设置采样率
    * @param rate 采样率(每秒帧数)
    */
-  setSampleRate(rate: number): void {
+  setSampleRate (rate: number): void {
     this.sampleRate = Math.max(1, rate);
   }
 
   /**
    * 获取采样率
    */
-  getSampleRate(): number {
+  getSampleRate (): number {
     return this.sampleRate;
   }
 
@@ -177,9 +181,9 @@ export class AnimationClip extends ReferResource {
    * @param time 时间(秒)
    * @returns 采样结果，每个通道的当前值 { trackName: value }
    */
-  sample(time: number): Record<string, any> {
+  sample (time: number): Record<string, any> {
     const result: Record<string, any> = {};
-    
+
     // 如果启用循环，处理时间循环
     if (this.loop && this.duration > 0) {
       time = time % this.duration;
@@ -187,15 +191,16 @@ export class AnimationClip extends ReferResource {
         time += this.duration;
       }
     }
-    
+
     // 对每个通道进行采样
     for (const track of this.tracks) {
       const value = this.sampleTrack(track, time);
+
       if (value !== undefined) {
         result[track.path] = value;
       }
     }
-    
+
     return result;
   }
 
@@ -207,49 +212,48 @@ export class AnimationClip extends ReferResource {
    */
   private sampleTrack<T>(track: AnimationTrack<T>, time: number): T | undefined {
     const keyframes = track.keyframes;
-    
+
     if (keyframes.length === 0) {
       return undefined;
     }
-    
+
     // 如果只有一个关键帧或时间在第一个关键帧之前
     if (keyframes.length === 1 || time <= keyframes[0].time) {
       return keyframes[0].value;
     }
-    
+
     // 如果时间在最后一个关键帧之后
     if (time >= keyframes[keyframes.length - 1].time) {
       return keyframes[keyframes.length - 1].value;
     }
-    
+
     // 查找时间所在的两个关键帧
     let i = 1;
+
     while (i < keyframes.length && time > keyframes[i].time) {
       i++;
     }
-    
+
     const previousKeyframe = keyframes[i - 1];
     const nextKeyframe = keyframes[i];
-    
+
     // 计算两个关键帧之间的插值因子 (0-1)
     const t = (time - previousKeyframe.time) / (nextKeyframe.time - previousKeyframe.time);
-    
+
     // 根据插值类型计算当前值
     const interpolationType = nextKeyframe.interpolation || track.interpolation;
-    
+
     switch (interpolationType) {
       case InterpolationType.Step:
         // 阶梯插值，直接返回前一个关键帧的值
         return previousKeyframe.value;
-        
       case InterpolationType.Linear:
         // 线性插值
         return this.linearInterpolate(previousKeyframe.value, nextKeyframe.value, t);
-        
       case InterpolationType.CubicSpline:
         // 三次样条插值，需要切线信息
         if (
-          previousKeyframe.outTangent !== undefined && 
+          previousKeyframe.outTangent !== undefined &&
           nextKeyframe.inTangent !== undefined
         ) {
           return this.cubicInterpolate(
@@ -263,7 +267,6 @@ export class AnimationClip extends ReferResource {
           // 如果没有切线信息，退化为线性插值
           return this.linearInterpolate(previousKeyframe.value, nextKeyframe.value, t);
         }
-        
       default:
         return this.linearInterpolate(previousKeyframe.value, nextKeyframe.value, t);
     }
@@ -276,21 +279,21 @@ export class AnimationClip extends ReferResource {
    * @param t 插值因子(0-1)
    * @returns 插值结果
    */
-  private linearInterpolate(a: any, b: any, t: number): any {
+  private linearInterpolate (a: any, b: any, t: number): any {
     // 处理数字
     if (typeof a === 'number' && typeof b === 'number') {
       return a + (b - a) * t;
     }
-    
+
     // 处理数组(如向量)
     if (Array.isArray(a) && Array.isArray(b) && a.length === b.length) {
       return a.map((value, index) => value + (b[index] - value) * t);
     }
-    
+
     // 处理对象
     if (typeof a === 'object' && typeof b === 'object' && a !== null && b !== null) {
       const result: any = {};
-      
+
       // 只插值两个对象都有的属性
       for (const key in a) {
         if (key in b) {
@@ -299,10 +302,10 @@ export class AnimationClip extends ReferResource {
           result[key] = a[key];
         }
       }
-      
+
       return result;
     }
-    
+
     // 默认情况下，不能插值，返回最接近的值
     return t < 0.5 ? a : b;
   }
@@ -316,34 +319,34 @@ export class AnimationClip extends ReferResource {
    * @param t 插值因子(0-1)
    * @returns 插值结果
    */
-  private cubicInterpolate(a: any, b: any, outTangent: any, inTangent: any, t: number): any {
+  private cubicInterpolate (a: any, b: any, outTangent: any, inTangent: any, t: number): any {
     // 处理数字
     if (typeof a === 'number' && typeof b === 'number' &&
         typeof outTangent === 'number' && typeof inTangent === 'number') {
       // 三次Hermite插值公式
       const t2 = t * t;
       const t3 = t2 * t;
-      
+
       const h00 = 2 * t3 - 3 * t2 + 1;
       const h10 = t3 - 2 * t2 + t;
       const h01 = -2 * t3 + 3 * t2;
       const h11 = t3 - t2;
-      
+
       return h00 * a + h10 * outTangent + h01 * b + h11 * inTangent;
     }
-    
+
     // 处理数组(如向量)
-    if (Array.isArray(a) && Array.isArray(b) && 
+    if (Array.isArray(a) && Array.isArray(b) &&
         Array.isArray(outTangent) && Array.isArray(inTangent) &&
         a.length === b.length && a.length === outTangent.length && a.length === inTangent.length) {
-      
-      return a.map((value, index) => 
+
+      return a.map((value, index) =>
         this.cubicInterpolate(
           value, b[index], outTangent[index], inTangent[index], t
         ) as number
       );
     }
-    
+
     // 对于不支持的类型，退化为线性插值
     return this.linearInterpolate(a, b, t);
   }
@@ -353,7 +356,7 @@ export class AnimationClip extends ReferResource {
    * @param target 目标对象
    * @returns 动画实例
    */
-  createInstance(target: object): AnimationInstance {
+  createInstance (target: object): AnimationInstance {
     return new AnimationInstance(this, target);
   }
 }
@@ -375,11 +378,11 @@ export class AnimationInstance extends Animation {
    * @param clip 动画片段
    * @param target 目标对象
    */
-  constructor(clip: AnimationClip, target: object) {
+  constructor (clip: AnimationClip, target: object) {
     super(clip.getName(), clip.getDuration());
     this.clip = clip;
     this.target = target;
-    
+
     // 默认循环设置
     this.setLoopMode(clip.isLoop() ? LoopMode.Loop : LoopMode.Once);
   }
@@ -387,14 +390,14 @@ export class AnimationInstance extends Animation {
   /**
    * 获取动画片段
    */
-  getClip(): AnimationClip {
+  getClip (): AnimationClip {
     return this.clip;
   }
 
   /**
    * 获取目标对象
    */
-  getTarget(): object {
+  getTarget (): object {
     return this.target;
   }
 
@@ -402,7 +405,7 @@ export class AnimationInstance extends Animation {
    * 设置目标对象
    * @param target 目标对象
    */
-  setTarget(target: object): void {
+  setTarget (target: object): void {
     this.target = target;
     // 重新采样确保更新到正确的目标
     this.sample();
@@ -411,10 +414,10 @@ export class AnimationInstance extends Animation {
   /**
    * 采样当前时间的动画帧
    */
-  protected override sample(): void {
+  protected override sample (): void {
     // 采样动画片段
     this.currentValues = this.clip.sample(this.getTime());
-    
+
     // 将采样结果应用到目标对象
     this.applyValues();
   }
@@ -422,11 +425,12 @@ export class AnimationInstance extends Animation {
   /**
    * 将采样的值应用到目标对象
    */
-  private applyValues(): void {
+  private applyValues (): void {
     if (!this.target) {return;}
-    
+
     for (const path in this.currentValues) {
       const value = this.currentValues[path];
+
       this.setNestedProperty(this.target, path, value);
     }
   }
@@ -437,31 +441,31 @@ export class AnimationInstance extends Animation {
    * @param path 属性路径 (如 "transform.position.x")
    * @param value 要设置的值
    */
-  private setNestedProperty(obj: any, path: string, value: any): void {
+  private setNestedProperty (obj: any, path: string, value: any): void {
     if (!obj || !path) {return;}
-    
+
     const parts = path.split('.');
     let current = obj;
-    
+
     // 遍历路径的每一部分，直到最后一个属性
     for (let i = 0; i < parts.length - 1; i++) {
       const part = parts[i];
-      
+
       // 如果当前属性是函数，尝试调用它
       if (typeof current[part] === 'function') {
         current = current[part]();
       } else {
         current = current[part];
       }
-      
+
       // 如果路径中间断了，停止处理
       if (current === undefined || current === null) {
         return;
       }
     }
-    
+
     const lastPart = parts[parts.length - 1];
-    
+
     // 处理特殊情况，如向量组件
     if (current[lastPart] !== undefined) {
       if (typeof current[lastPart] === 'function') {
@@ -476,4 +480,4 @@ export class AnimationInstance extends Animation {
       current.set(lastPart, value);
     }
   }
-} 
+}
