+++ packages/core/src/camera/camera.ts
@@ -1,5 +1,5 @@
 import { Component } from '../base/component';
-import { Entity } from '../base/entity';
+import type { Entity } from '../base/entity';
 import { Vector3, Matrix4 } from '@maxellabs/math';
 
 /**
@@ -51,17 +51,17 @@ export class Camera extends Component {
    * 创建一个新的摄像机组件
    * @param entity 所属实体
    */
-  constructor(entity: Entity) {
+  constructor (entity: Entity) {
     super(entity);
   }
 
   /** 获取视野角度（度数） */
-  get fov(): number {
+  get fov (): number {
     return this._fov;
   }
 
   /** 设置视野角度（度数） */
-  set fov(value: number) {
+  set fov (value: number) {
     if (this._fov !== value) {
       this._fov = value;
       this._projectionDirty = true;
@@ -69,12 +69,12 @@ export class Camera extends Component {
   }
 
   /** 获取近裁剪面距离 */
-  get near(): number {
+  get near (): number {
     return this._near;
   }
 
   /** 设置近裁剪面距离 */
-  set near(value: number) {
+  set near (value: number) {
     if (this._near !== value) {
       this._near = value;
       this._projectionDirty = true;
@@ -82,12 +82,12 @@ export class Camera extends Component {
   }
 
   /** 获取远裁剪面距离 */
-  get far(): number {
+  get far (): number {
     return this._far;
   }
 
   /** 设置远裁剪面距离 */
-  set far(value: number) {
+  set far (value: number) {
     if (this._far !== value) {
       this._far = value;
       this._projectionDirty = true;
@@ -95,12 +95,12 @@ export class Camera extends Component {
   }
 
   /** 获取宽高比 */
-  get aspect(): number {
+  get aspect (): number {
     return this._aspect;
   }
 
   /** 设置宽高比 */
-  set aspect(value: number) {
+  set aspect (value: number) {
     if (this._aspect !== value) {
       this._aspect = value;
       this._projectionDirty = true;
@@ -108,12 +108,12 @@ export class Camera extends Component {
   }
 
   /** 获取投影类型 */
-  get projectionType(): ProjectionType {
+  get projectionType (): ProjectionType {
     return this._projectionType;
   }
 
   /** 设置投影类型 */
-  set projectionType(value: ProjectionType) {
+  set projectionType (value: ProjectionType) {
     if (this._projectionType !== value) {
       this._projectionType = value;
       this._projectionDirty = true;
@@ -121,12 +121,12 @@ export class Camera extends Component {
   }
 
   /** 获取正交投影尺寸（高度的一半） */
-  get orthographicSize(): number {
+  get orthographicSize (): number {
     return this._orthographicSize;
   }
 
   /** 设置正交投影尺寸（高度的一半） */
-  set orthographicSize(value: number) {
+  set orthographicSize (value: number) {
     if (this._orthographicSize !== value) {
       this._orthographicSize = value;
       this._projectionDirty = true;
@@ -134,78 +134,80 @@ export class Camera extends Component {
   }
 
   /** 获取视口矩形（x, y, width, height） */
-  get viewport(): [number, number, number, number] {
+  get viewport (): [number, number, number, number] {
     return [...this._viewport];
   }
 
   /** 设置视口矩形（x, y, width, height） */
-  set viewport(value: [number, number, number, number]) {
+  set viewport (value: [number, number, number, number]) {
     this._viewport = [...value];
   }
 
   /** 获取渲染优先级（数值越小优先级越高） */
-  get priority(): number {
+  get priority (): number {
     return this._priority;
   }
 
   /** 设置渲染优先级（数值越小优先级越高） */
-  set priority(value: number) {
+  set priority (value: number) {
     this._priority = value;
   }
 
   /** 获取背景颜色 */
-  get backgroundColor(): [number, number, number, number] {
+  get backgroundColor (): [number, number, number, number] {
     return [...this._backgroundColor];
   }
 
   /** 设置背景颜色 */
-  set backgroundColor(value: [number, number, number, number]) {
+  set backgroundColor (value: [number, number, number, number]) {
     this._backgroundColor = [...value];
   }
 
   /** 获取是否清除深度缓冲 */
-  get clearDepth(): boolean {
+  get clearDepth (): boolean {
     return this._clearDepth;
   }
 
   /** 设置是否清除深度缓冲 */
-  set clearDepth(value: boolean) {
+  set clearDepth (value: boolean) {
     this._clearDepth = value;
   }
 
   /** 获取是否清除颜色缓冲 */
-  get clearColor(): boolean {
+  get clearColor (): boolean {
     return this._clearColor;
   }
 
   /** 设置是否清除颜色缓冲 */
-  set clearColor(value: boolean) {
+  set clearColor (value: boolean) {
     this._clearColor = value;
   }
 
   /** 获取投影矩阵 */
-  get projectionMatrix(): Matrix4 {
+  get projectionMatrix (): Matrix4 {
     if (this._projectionDirty) {
       this._updateProjectionMatrix();
     }
+
     return this._projectionMatrix.clone();
   }
 
   /** 获取视图矩阵 */
-  get viewMatrix(): Matrix4 {
+  get viewMatrix (): Matrix4 {
     if (this._viewDirty) {
       this._updateViewMatrix();
     }
+
     return this._viewMatrix.clone();
   }
 
   /** 获取视图投影矩阵（VP矩阵） */
-  get viewProjectionMatrix(): Matrix4 {
-    return Matrix4.multiply(this.projectionMatrix, this.viewMatrix);
+  get viewProjectionMatrix (): Matrix4 {
+    return this.projectionMatrix.clone().multiply(this.viewMatrix);
   }
 
   /** 更新投影矩阵 */
-  private _updateProjectionMatrix(): void {
+  private _updateProjectionMatrix (): void {
     if (this._projectionType === ProjectionType.Perspective) {
       // 透视投影
       this._projectionMatrix = new Matrix4().perspective(
@@ -218,6 +220,7 @@ export class Camera extends Component {
       // 正交投影
       const height = this._orthographicSize * 2;
       const width = height * this._aspect;
+
       this._projectionMatrix = new Matrix4().orthographic(
         -width / 2,
         width / 2,
@@ -231,7 +234,7 @@ export class Camera extends Component {
   }
 
   /** 更新视图矩阵 */
-  private _updateViewMatrix(): void {
+  private _updateViewMatrix (): void {
     // 视图矩阵是相机世界矩阵的逆矩阵
     this._viewMatrix = Matrix4.invert(this.entity.transform.worldMatrix);
     this._viewDirty = false;
@@ -244,17 +247,17 @@ export class Camera extends Component {
    * @param screenHeight 屏幕高度
    * @returns 屏幕坐标点[x, y]，x和y范围都是0到1
    */
-  worldToScreenPoint(worldPosition: Vector3, screenWidth: number, screenHeight: number): [number, number] {
+  worldToScreenPoint (worldPosition: Vector3, screenWidth: number, screenHeight: number): [number, number] {
     // 获取视图投影矩阵
     const vpMatrix = this.viewProjectionMatrix;
-    
+
     // 将世界坐标转换为裁剪空间坐标
     const clipSpacePos = vpMatrix.transformPoint(worldPosition);
-    
+
     // 裁剪空间坐标范围是 [-1, 1]，将其转换为 [0, 1] 范围的屏幕坐标
     const screenX = (clipSpacePos.x + 1) * 0.5;
     const screenY = (1 - clipSpacePos.y) * 0.5; // Y轴需要翻转
-    
+
     return [screenX, screenY];
   }
 
@@ -264,35 +267,35 @@ export class Camera extends Component {
    * @param screenY 屏幕Y坐标（0到1）
    * @returns 从相机位置指向屏幕点的射线方向
    */
-  screenPointToRay(screenX: number, screenY: number): { origin: Vector3, direction: Vector3 } {
+  screenPointToRay (screenX: number, screenY: number): { origin: Vector3, direction: Vector3 } {
     // 将屏幕坐标转换为裁剪空间坐标（-1到1）
     const clipX = screenX * 2 - 1;
     const clipY = 1 - screenY * 2; // Y轴需要翻转
-    
+
     // 创建近平面和远平面上的点
     const nearPlanePoint = new Vector3(clipX, clipY, -1);
     const farPlanePoint = new Vector3(clipX, clipY, 1);
-    
+
     // 将视图投影矩阵求逆
     const vpMatrixInverse = Matrix4.invert(this.viewProjectionMatrix);
-    
+
     // 将裁剪空间坐标转换为世界坐标
     const nearWorldPoint = vpMatrixInverse.transformPoint(nearPlanePoint);
     const farWorldPoint = vpMatrixInverse.transformPoint(farPlanePoint);
-    
+
     // 射线原点（相机位置）
     const origin = this.entity.transform.worldPosition;
-    
+
     // 射线方向
     const direction = Vector3.subtract(farWorldPoint, nearWorldPoint).normalize();
-    
+
     return { origin, direction };
   }
 
   /**
    * 当组件启用时调用
    */
-  onEnable(): void {
+  onEnable (): void {
     // 如果是场景中第一个启用的相机，将其设置为主相机
     if (this.entity.scene && !this.entity.scene.mainCamera) {
       this.entity.scene.mainCamera = this.entity;
@@ -302,9 +305,10 @@ export class Camera extends Component {
   /**
    * 在每一帧更新前调用
    */
-  update(deltaTime: number): void {
+  update (deltaTime: number): void {
     // 检查相机变换是否发生变化
     const transform = this.entity.transform;
+
     if (transform.worldMatrix.isDirty()) {
       this._viewDirty = true;
     }
@@ -313,8 +317,8 @@ export class Camera extends Component {
   /**
    * 在渲染时调用
    */
-  render(): void {
+  render (): void {
     // 实际渲染操作通常由渲染系统处理
     // 这里只是一个占位符
   }
-} 
+}
