+++ packages/core/src/serialization/Serialization.ts
@@ -23,30 +23,30 @@ export enum SerializeFlags {
  */
 export interface SerializeField {
   /** 序列化名称，默认使用属性名 */
-  name?: string;
+  name?: string,
   /** 是否可序列化，默认true */
-  serializable?: boolean;
+  serializable?: boolean,
   /** 默认值 */
-  defaultValue?: any;
+  defaultValue?: any,
   /** 自定义序列化函数 */
-  serialize?: (value: any) => any;
+  serialize?: (value: any) => any,
   /** 自定义反序列化函数 */
-  deserialize?: (data: any) => any;
+  deserialize?: (data: any) => any,
   /** 类型信息 */
-  type?: any;
+  type?: any,
 }
 
 /**
  * 序列化字段元数据
  */
 interface SerializeFieldInfo {
-  name: string;
-  propertyKey: string;
-  defaultValue?: any;
-  serializable: boolean;
-  serialize?: (value: any) => any;
-  deserialize?: (data: any) => any;
-  type?: any;
+  name: string,
+  propertyKey: string,
+  defaultValue?: any,
+  serializable: boolean,
+  serialize?: (value: any) => any,
+  deserialize?: (data: any) => any,
+  type?: any,
 }
 
 /**
@@ -57,21 +57,21 @@ export interface ISerializable {
    * 自定义序列化方法
    * @returns 序列化后的对象
    */
-  _serialize(): any;
-  
+  _serialize(): any,
+
   /**
    * 自定义反序列化方法
    * @param data 反序列化数据
    */
-  _deserialize(data: any): void;
+  _deserialize(data: any): void,
 }
 
 /**
  * 类型信息
  */
 interface TypeInfo {
-  type: string;
-  module?: string;
+  type: string,
+  module?: string,
 }
 
 /**
@@ -88,21 +88,21 @@ const typeRegistry: Map<string, Function> = new Map();
  * 序列化字段装饰器
  * @param options 字段配置选项
  */
-export function serializable(options: SerializeField = {}): PropertyDecorator {
-  return function(target: any, propertyKey: string | symbol) {
+export function serializable (options: SerializeField = {}): PropertyDecorator {
+  return function (target: any, propertyKey: string | symbol) {
     if (typeof propertyKey !== 'string') {
       return;
     }
-    
+
     const constructor = target.constructor;
-    
+
     // 确保类有元数据映射
     if (!metadataMap.has(constructor)) {
       metadataMap.set(constructor, new Map());
     }
-    
+
     const metadata = metadataMap.get(constructor);
-    
+
     // 保存字段元数据
     metadata.set(propertyKey, {
       name: options.name || propertyKey,
@@ -111,7 +111,7 @@ export function serializable(options: SerializeField = {}): PropertyDecorator {
       serializable: options.serializable !== false,
       serialize: options.serialize,
       deserialize: options.deserialize,
-      type: options.type
+      type: options.type,
     });
   };
 }
@@ -121,12 +121,13 @@ export function serializable(options: SerializeField = {}): PropertyDecorator {
  * @param typeName 类型名称
  * @param module 模块名
  */
-export function serializableClass(typeName?: string, module?: string): ClassDecorator {
-  return function(target: Function) {
+export function serializableClass (typeName?: string, module?: string): ClassDecorator {
+  return function (target: Function) {
     // 注册类型
     const name = typeName || target.name;
+
     typeRegistry.set(module ? `${module}.${name}` : name, target);
-    
+
     // 确保有元数据映射
     if (!metadataMap.has(target)) {
       metadataMap.set(target, new Map());
@@ -144,136 +145,141 @@ export class SerializationManager {
    * @param typeName 类型名称
    * @param module 模块名
    */
-  static registerType(type: Function, typeName?: string, module?: string): void {
+  static registerType (type: Function, typeName?: string, module?: string): void {
     const name = typeName || type.name;
+
     typeRegistry.set(module ? `${module}.${name}` : name, type);
   }
-  
+
   /**
    * 获取类型
    * @param typeName 类型名称
    * @returns 类型构造函数
    */
-  static getType(typeName: string): Function | undefined {
+  static getType (typeName: string): Function | undefined {
     return typeRegistry.get(typeName);
   }
-  
+
   /**
    * 对象序列化
    * @param obj 待序列化对象
    * @param flags 序列化标志
    * @returns 序列化后的JSON字符串
    */
-  static serialize(obj: any, flags: SerializeFlags = SerializeFlags.Default): string {
+  static serialize (obj: any, flags: SerializeFlags = SerializeFlags.Default): string {
     const data = this.toJSON(obj, flags);
-    
+
     const indent = flags & SerializeFlags.PrettyPrint ? 2 : undefined;
+
     return JSON.stringify(data, null, indent);
   }
-  
+
   /**
    * 将对象转换为可序列化的JSON对象
    * @param obj 源对象
    * @param flags 序列化标志
    * @returns 可序列化的JSON对象
    */
-  static toJSON(obj: any, flags: SerializeFlags = SerializeFlags.Default): any {
+  static toJSON (obj: any, flags: SerializeFlags = SerializeFlags.Default): any {
     if (obj === null || obj === undefined) {
       return obj;
     }
-    
+
     // 处理原始类型
     if (typeof obj !== 'object') {
       return obj;
     }
-    
+
     // 处理数组
     if (Array.isArray(obj)) {
       return obj.map(item => this.toJSON(item, flags));
     }
-    
+
     // 处理Date对象
     if (obj instanceof Date) {
       return {
         __type: 'Date',
-        value: obj.toISOString()
+        value: obj.toISOString(),
       };
     }
-    
+
     // 处理Map对象
     if (obj instanceof Map) {
       const entries = Array.from(obj.entries()).map(([key, value]) => ({
         key: this.toJSON(key, flags),
-        value: this.toJSON(value, flags)
+        value: this.toJSON(value, flags),
       }));
-      
+
       return {
         __type: 'Map',
-        entries
+        entries,
       };
     }
-    
+
     // 处理Set对象
     if (obj instanceof Set) {
       return {
         __type: 'Set',
-        values: Array.from(obj).map(item => this.toJSON(item, flags))
+        values: Array.from(obj).map(item => this.toJSON(item, flags)),
       };
     }
-    
+
     // 使用自定义序列化方法
     if (typeof obj._serialize === 'function') {
       const data = obj._serialize();
-      
+
       // 添加类型信息
       if (flags & SerializeFlags.IncludeTypeInfo) {
         for (const [constructor, typeInfo] of typeRegistry.entries()) {
           if (obj instanceof constructor) {
             data.__type = constructor.name;
+
             break;
           }
         }
       }
-      
+
       return data;
     }
-    
+
     // 创建结果对象
     const result: any = {};
-    
+
     // 添加类型信息
     if (flags & SerializeFlags.IncludeTypeInfo) {
       const constructor = obj.constructor;
+
       if (constructor && constructor !== Object) {
         for (const [typeName, typeConstructor] of typeRegistry.entries()) {
           if (constructor === typeConstructor) {
             result.__type = typeName;
+
             break;
           }
         }
       }
     }
-    
+
     // 获取类的元数据
     const metadata = metadataMap.get(obj.constructor) || new Map();
-    
+
     // 序列化所有字段
     for (const [propertyKey, fieldInfo] of metadata.entries()) {
-      if (!fieldInfo.serializable) continue;
-      
+      if (!fieldInfo.serializable) {continue;}
+
       const value = obj[propertyKey];
-      
+
       // 忽略空值
       if ((flags & SerializeFlags.IgnoreNull && value === null) ||
           (flags & SerializeFlags.IgnoreUndefined && value === undefined)) {
         continue;
       }
-      
+
       // 忽略默认值
       if (flags & SerializeFlags.IgnoreDefault && value === fieldInfo.defaultValue) {
         continue;
       }
-      
+
       // 使用自定义序列化函数
       if (fieldInfo.serialize) {
         result[fieldInfo.name] = fieldInfo.serialize(value);
@@ -281,7 +287,7 @@ export class SerializationManager {
         result[fieldInfo.name] = this.toJSON(value, flags);
       }
     }
-    
+
     // 处理没有元数据的普通对象属性
     if (metadata.size === 0) {
       for (const key in obj) {
@@ -290,23 +296,23 @@ export class SerializationManager {
           if (flags & SerializeFlags.IgnorePrivate && key.startsWith('_')) {
             continue;
           }
-          
+
           const value = obj[key];
-          
+
           // 忽略空值
           if ((flags & SerializeFlags.IgnoreNull && value === null) ||
               (flags & SerializeFlags.IgnoreUndefined && value === undefined)) {
             continue;
           }
-          
+
           result[key] = this.toJSON(value, flags);
         }
       }
     }
-    
+
     return result;
   }
-  
+
   /**
    * 反序列化对象
    * @param json JSON字符串或对象
@@ -315,9 +321,10 @@ export class SerializationManager {
    */
   static deserialize<T>(json: string | any, targetType?: Function): T {
     const data = typeof json === 'string' ? JSON.parse(json) : json;
+
     return this.fromJSON<T>(data, targetType);
   }
-  
+
   /**
    * 从JSON对象创建实例
    * @param data JSON数据
@@ -326,70 +333,76 @@ export class SerializationManager {
    */
   static fromJSON<T>(data: any, targetType?: Function): T {
     if (data === null || data === undefined) {
-      return data as any;
+      return data;
     }
-    
+
     // 处理原始类型
     if (typeof data !== 'object') {
-      return data as any;
+      return data;
     }
-    
+
     // 处理数组
     if (Array.isArray(data)) {
       return data.map(item => this.fromJSON(item)) as any;
     }
-    
+
     // 处理特殊类型
     if (data.__type) {
       // 处理Date
       if (data.__type === 'Date') {
         return new Date(data.value) as any;
       }
-      
+
       // 处理Map
       if (data.__type === 'Map') {
         const map = new Map();
+
         for (const entry of data.entries) {
           map.set(
             this.fromJSON(entry.key),
             this.fromJSON(entry.value)
           );
         }
+
         return map as any;
       }
-      
+
       // 处理Set
       if (data.__type === 'Set') {
         const set = new Set();
+
         for (const value of data.values) {
           set.add(this.fromJSON(value));
         }
+
         return set as any;
       }
-      
+
       // 处理自定义类型
       const typeConstructor = this.getType(data.__type) || targetType;
+
       if (typeConstructor) {
         return this.createInstanceFromData(typeConstructor, data);
       }
     }
-    
+
     // 如果提供了目标类型
     if (targetType) {
       return this.createInstanceFromData(targetType, data);
     }
-    
+
     // 处理普通对象
     const result: any = {};
+
     for (const key in data) {
       if (data.hasOwnProperty(key) && key !== '__type') {
         result[key] = this.fromJSON(data[key]);
       }
     }
-    
+
     return result as T;
   }
-  
+
   /**
    * 从数据创建类型实例
    * @param type 类型构造函数
@@ -399,25 +412,26 @@ export class SerializationManager {
   private static createInstanceFromData<T>(type: Function, data: any): T {
     // 创建实例
     const instance = new (type as any)();
-    
+
     // 使用自定义反序列化方法
     if (typeof instance._deserialize === 'function') {
       instance._deserialize(data);
+
       return instance;
     }
-    
+
     // 获取类的元数据
     const metadata = metadataMap.get(type) || new Map();
-    
+
     // 处理所有可序列化字段
     for (const [propertyKey, fieldInfo] of metadata.entries()) {
-      if (!fieldInfo.serializable) continue;
-      
+      if (!fieldInfo.serializable) {continue;}
+
       const dataKey = fieldInfo.name;
-      
+
       if (dataKey in data) {
         const value = data[dataKey];
-        
+
         // 使用自定义反序列化函数
         if (fieldInfo.deserialize) {
           instance[propertyKey] = fieldInfo.deserialize(value);
@@ -433,7 +447,7 @@ export class SerializationManager {
         instance[propertyKey] = fieldInfo.defaultValue;
       }
     }
-    
+
     // 如果没有元数据，处理所有属性
     if (metadata.size === 0) {
       for (const key in data) {
@@ -442,10 +456,10 @@ export class SerializationManager {
         }
       }
     }
-    
+
     return instance as T;
   }
-  
+
   /**
    * 克隆对象
    * @param obj 源对象
@@ -453,6 +467,7 @@ export class SerializationManager {
    */
   static clone<T>(obj: T): T {
     const json = this.serialize(obj, SerializeFlags.IncludeTypeInfo);
+
     return this.deserialize(json);
   }
-} 
+}
