+++ packages/core/src/shader/ShaderProperty.ts
@@ -47,7 +47,7 @@ export class ShaderProperty {
    * @param nameOrID 属性名称或ID
    * @param type 属性类型
    */
-  constructor(nameOrID: string | number, type: ShaderPropertyType = ShaderPropertyType.Float) {
+  constructor (nameOrID: string | number, type: ShaderPropertyType = ShaderPropertyType.Float) {
     if (typeof nameOrID === 'string') {
       this._name = nameOrID;
       this._id = ShaderProperty.getIDByName(nameOrID);
@@ -61,21 +61,21 @@ export class ShaderProperty {
   /**
    * 获取属性ID
    */
-  get id(): number {
+  get id (): number {
     return this._id;
   }
 
   /**
    * 获取属性名称
    */
-  get name(): string {
+  get name (): string {
     return this._name;
   }
 
   /**
    * 获取属性类型
    */
-  get type(): ShaderPropertyType {
+  get type (): ShaderPropertyType {
     return this._type;
   }
 
@@ -84,14 +84,16 @@ export class ShaderProperty {
    * @param name 属性名称
    * @returns 属性ID
    */
-  static getIDByName(name: string): number {
+  static getIDByName (name: string): number {
     if (ShaderProperty._nameToIDMap.has(name)) {
       return ShaderProperty._nameToIDMap.get(name);
     }
 
     const id = ShaderProperty._propertyIDCounter++;
+
     ShaderProperty._nameToIDMap.set(name, id);
     ShaderProperty._idToNameMap.set(id, name);
+
     return id;
   }
 
@@ -100,7 +102,7 @@ export class ShaderProperty {
    * @param id 属性ID
    * @returns 属性名称
    */
-  static getNameById(id: number): string {
+  static getNameById (id: number): string {
     return ShaderProperty._idToNameMap.get(id) || '';
   }
 
@@ -109,7 +111,7 @@ export class ShaderProperty {
    * @param id 属性ID
    * @returns 是否存在
    */
-  static hasID(id: number): boolean {
+  static hasID (id: number): boolean {
     return ShaderProperty._idToNameMap.has(id);
   }
 
@@ -118,7 +120,7 @@ export class ShaderProperty {
    * @param name 属性名称
    * @returns 是否存在
    */
-  static hasName(name: string): boolean {
+  static hasName (name: string): boolean {
     return ShaderProperty._nameToIDMap.has(name);
   }
 
@@ -126,14 +128,14 @@ export class ShaderProperty {
    * 获取所有属性名称
    * @returns 属性名称数组
    */
-  static getAllPropertyNames(): string[] {
+  static getAllPropertyNames (): string[] {
     return Array.from(ShaderProperty._nameToIDMap.keys());
   }
 
   /**
    * 创建常用的内置着色器属性
    */
-  static createBuiltins(): void {
+  static createBuiltins (): void {
     // 创建常用的内置着色器属性
     // 矩阵
     ShaderProperty.getIDByName('modelMatrix');
@@ -142,7 +144,7 @@ export class ShaderProperty {
     ShaderProperty.getIDByName('viewProjectionMatrix');
     ShaderProperty.getIDByName('modelViewMatrix');
     ShaderProperty.getIDByName('normalMatrix');
-    
+
     // 常用纹理名称
     ShaderProperty.getIDByName('mainTexture');
     ShaderProperty.getIDByName('diffuseTexture');
@@ -152,7 +154,7 @@ export class ShaderProperty {
     ShaderProperty.getIDByName('roughnessTexture');
     ShaderProperty.getIDByName('metallicTexture');
     ShaderProperty.getIDByName('occlusionTexture');
-    
+
     // 常用材质参数
     ShaderProperty.getIDByName('diffuseColor');
     ShaderProperty.getIDByName('specularColor');
@@ -160,13 +162,13 @@ export class ShaderProperty {
     ShaderProperty.getIDByName('roughness');
     ShaderProperty.getIDByName('metallic');
     ShaderProperty.getIDByName('opacity');
-    
+
     // 常用相机参数
     ShaderProperty.getIDByName('cameraPosition');
     ShaderProperty.getIDByName('cameraDirection');
     ShaderProperty.getIDByName('nearPlane');
     ShaderProperty.getIDByName('farPlane');
-    
+
     // 常用光照参数
     ShaderProperty.getIDByName('lightPosition');
     ShaderProperty.getIDByName('lightDirection');
@@ -174,4 +176,4 @@ export class ShaderProperty {
     ShaderProperty.getIDByName('lightIntensity');
     ShaderProperty.getIDByName('ambientColor');
   }
-} 
+}
